+++ packages/@ionic/cli-utils/src/lib/build.ts
@@ -2,7 +2,7 @@ import chalk from 'chalk';
 
 import { BaseError } from '@ionic/cli-framework/lib/errors';
 
-import { BaseBuildOptions, BuildOptions, CommandLineInputs, CommandLineOptions, CommandMetadata, CommandMetadataOption, IConfig, IProject, IShell, IonicEnvironment, ProjectType } from '../definitions';
+import { BaseBuildOptions, BuildOptions, CommandLineInputs, CommandLineOptions, CommandMetadata, CommandMetadataOption, IConfig, ILogger, IProject, IShell, IonicEnvironment, ProjectType } from '../definitions';
 import { PROJECT_FILE } from '../constants';
 import { FatalException, RunnerException, RunnerNotFoundException } from './errors';
 import { Runner } from './runner';
@@ -28,18 +28,21 @@ export const COMMON_BUILD_COMMAND_OPTIONS: ReadonlyArray<CommandMetadataOption> 
 
 export interface BuildRunnerDeps {
   readonly config: IConfig;
+  readonly log: ILogger;
   readonly project: IProject;
   readonly shell: IShell;
 }
 
 export abstract class BuildRunner<T extends BuildOptions<any>> extends Runner<T, void> {
   protected readonly config: IConfig;
+  protected readonly log: ILogger;
   protected readonly project: IProject;
   protected readonly shell: IShell;
 
-  constructor({ config, project, shell }: BuildRunnerDeps) {
+  constructor({ config, log, project, shell }: BuildRunnerDeps) {
     super();
     this.config = config;
+    this.log = log;
     this.project = project;
     this.shell = shell;
   }
@@ -94,6 +97,10 @@ export abstract class BuildRunner<T extends BuildOptions<any>> extends Runner<T,
   }
 
   async run(options: T): Promise<void> {
+    if (options.engine === 'cordova' && !options.platform) {
+      this.log.warn(`Cordova engine chosen without a target platform. This could cause issues. Please use the ${chalk.green('--platform')} option.`);
+    }
+
     await this.beforeBuild(options);
     await this.buildProject(options);
     await this.afterBuild(options);